% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.priorityelasticnet.R
\name{coef.priorityelasticnet}
\alias{coef.priorityelasticnet}
\title{Extract coefficients from a priorityelasticnet object}
\usage{
\method{coef}{priorityelasticnet}(object, ...)
}
\arguments{
\item{object}{model of type priorityelasticnet}

\item{...}{additional arguments, currently not used}
}
\value{
List with the coefficients and the intercepts
}
\description{
Extract coefficients from a priorityelasticnet object
}
