% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pen_Data}
\alias{Pen_Data}
\title{Simulated Patient Data for Binary Classification}
\format{
A data frame with 406 rows and 325 columns:
\describe{
  \item{Clinical_Var1}{Numeric variable representing age.}
  \item{Clinical_Var2}{Binary variable representing gender (0 = male, 1 = female).}
  \item{Clinical_Var3}{Categorical variable representing race (values 0, 1, 2, or 3).}
  \item{Clinical_Var4}{Binary variable representing ethnicity (0 or 1).}
  \item{Clinical_Var5}{Binary variable representing radiation therapy status (0 or 1).}
  \item{Proteomic_Var1}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var2}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var3}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var4}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var5}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var6}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var7}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var8}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var9}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var10}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var11}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var12}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var13}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var14}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var15}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var16}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var17}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var18}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var19}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var20}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var21}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var22}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var23}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var24}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var25}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var26}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var27}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var28}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var29}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var30}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var31}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var32}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var33}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var34}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var35}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var36}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var37}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var38}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var39}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var40}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var41}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var42}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var43}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var44}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var45}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var46}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var47}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var48}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var49}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var50}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var51}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var52}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var53}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var54}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var55}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var56}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var57}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var58}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var59}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var60}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var61}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var62}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var63}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var64}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var65}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var66}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var67}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var68}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var69}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var70}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var71}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var72}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var73}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var74}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var75}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var76}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var77}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var78}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var79}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var80}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var81}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var82}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var83}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var84}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var85}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var86}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var87}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var88}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var89}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var90}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var91}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var92}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var93}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var94}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var95}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var96}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var97}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var98}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var99}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var100}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var101}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var102}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var103}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var104}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var105}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var106}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var107}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var108}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var109}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var110}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var111}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var112}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var113}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var114}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var115}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var116}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var117}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var118}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var119}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var120}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var121}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var122}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var123}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var124}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var125}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var126}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var127}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var128}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var129}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var130}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var131}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var132}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var133}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var134}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var135}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var136}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var137}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var138}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var139}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var140}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var141}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var142}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var143}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var144}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var145}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var146}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var147}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var148}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var149}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var150}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var151}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var152}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var153}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var154}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var155}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var156}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var157}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var158}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var159}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var160}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var161}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var162}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var163}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var164}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var165}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var166}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var167}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var168}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var169}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var170}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var171}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var172}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var173}{Continuous variable representing a proteomic measurement.}
  \item{Proteomic_Var174}{Continuous variable representing a proteomic measurement.}
  \item{RNA_Var1}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var2}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var3}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var4}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var5}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var6}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var7}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var8}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var9}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var10}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var11}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var12}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var13}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var14}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var15}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var16}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var17}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var18}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var19}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var20}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var21}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var22}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var23}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var24}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var25}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var26}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var27}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var28}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var29}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var30}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var31}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var32}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var33}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var34}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var35}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var36}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var37}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var38}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var39}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var40}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var41}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var42}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var43}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var44}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var45}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var46}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var47}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var48}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var49}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var50}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var51}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var52}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var53}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var54}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var55}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var56}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var57}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var58}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var59}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var60}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var61}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var62}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var63}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var64}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var65}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var66}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var67}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var68}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var69}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var70}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var71}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var72}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var73}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var74}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var75}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var76}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var77}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var78}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var79}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var80}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var81}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var82}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var83}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var84}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var85}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var86}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var87}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var88}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var89}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var90}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var91}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var92}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var93}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var94}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var95}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var96}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var97}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var98}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var99}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var100}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var101}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var102}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var103}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var104}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var105}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var106}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var107}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var108}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var109}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var110}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var111}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var112}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var113}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var114}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var115}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var116}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var117}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var118}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var119}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var120}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var121}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var122}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var123}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var124}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var125}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var126}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var127}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var128}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var129}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var130}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var131}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var132}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var133}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var134}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var135}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var136}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var137}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var138}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var139}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var140}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var141}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var142}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var143}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var144}{Continuous variable representing an RNA measurement.}
  \item{RNA_Var145}{Continuous variable representing an RNA measurement.}
  \item{Pen_out}{Binary outcome variable generated using a logistic function applied to a linear predictor based on the combined variables.}
}
}
\usage{
Pen_Data
}
\description{
This dataset contains simulated data for a binary classification problem, representing patient data with clinical, proteomics, and RNA variables. The data is organized into three blocks of variables: clinical variables, proteomics variables, and RNA variables. The outcome is a binary variable generated based on a logistic function.
}
\keyword{datasets}
